refactor(config): split datadog witness domains
