+++ lib/agent-data-plane-config-system/src/witness_impl/workload.rs
@@ -1,0 +1,19 @@
+use serde_json::Value;
+
+use super::*;
+
+pub(super) fn consume_key(translator: &mut Translator, key: &str, value: Value) -> Option<Value> {
+    match key {
+        "cri_connection_timeout" => {
+            translator.native.components.workload.source.cri_connection_timeout_secs = u64_value(value, 0);
+
+            None
+        }
+        "cri_query_timeout" => {
+            translator.native.components.workload.source.cri_query_timeout_secs = u64_value(value, 0);
+
+            None
+        }
+        _ => Some(value),
+    }
+}
